% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{ant}
\alias{ant}
\title{ant}
\usage{
ant(
  geom,
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xend = NULL,
  yend = NULL,
  ...,
  na.rm = FALSE
)
}
\description{
See \code{ggplot2::\link[ggplot2:annotate]{annotate}} for details.
}
\keyword{internal}
