% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{as_tsbl}
\alias{as_tsbl}
\title{as_tsbl}
\usage{
as_tsbl(
  x,
  key = NULL,
  index,
  regular = TRUE,
  validate = TRUE,
  .drop = TRUE,
  ...
)
}
\description{
See \code{tsibble::\link[tsibble::as_tsibble]{as_tsibble}} for details.
}
\keyword{internal}
