% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{ccr}
\alias{ccr}
\title{ccr}
\usage{
ccr(xlim = NULL, ylim = NULL, expand = TRUE, default = FALSE, clip = "on")
}
\description{
See \code{ggplot2::\link[ggplot2:coord_cartesian]{coord_cartesian}} for details.
}
\keyword{internal}
