% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{cfl}
\alias{cfl}
\title{cfl}
\usage{
cfl(xlim = NULL, ylim = NULL, expand = TRUE, clip = "on")
}
\description{
See \code{ggplot2::\link[ggplot2:coord_flip]{coord_flip}} for details.
}
\keyword{internal}
