% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{eln}
\alias{eln}
\title{eln}
\usage{
eln(
  colour = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE
)
}
\description{
See \code{ggplot2::\link[ggplot2:element_line]{element_line}} for details.
}
\keyword{internal}
