% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{erc}
\alias{erc}
\title{erc}
\usage{
erc(
  fill = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE
)
}
\description{
See \code{ggplot2::\link[ggplot2:element_rect]{element_rect}} for details.
}
\keyword{internal}
