% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{etx}
\alias{etx}
\title{etx}
\usage{
etx(
  family = NULL,
  face = NULL,
  colour = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  color = NULL,
  margin = NULL,
  debug = NULL,
  inherit.blank = FALSE
)
}
\description{
See \code{ggplot2::\link[ggplot2:element_text]{element_text}} for details.
}
\keyword{internal}
