% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{fgr}
\alias{fgr}
\title{fgr}
\usage{
fgr(
  rows = NULL,
  cols = NULL,
  scales = "fixed",
  space = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = TRUE,
  margins = FALSE,
  facets = NULL
)
}
\description{
See \code{ggplot2::\link[ggplot2:facet_grid]{facet_grid}} for details.
}
\keyword{internal}
