% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{fwr}
\alias{fwr}
\title{fwr}
\usage{
fwr(
  facets,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = TRUE,
  dir = "h",
  strip.position = "top"
)
}
\description{
See \code{ggplot2::\link[ggplot2:facet_wrap]{facet_wrap}} for details.
}
\keyword{internal}
