% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{gat}
\alias{gat}
\title{gat}
\usage{
gat(data, key = "key", value = "value", ..., na.rm = FALSE,
  convert = FALSE, factor_key = FALSE)
}
\description{
See \code{tidyr::\link[tidyr:gather]{gather}} for details.
}
\keyword{internal}
