% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{gby}
\alias{gby}
\title{gby}
\usage{
gby(.data, ..., add = FALSE, .drop = group_by_drop_default(.data))
}
\description{
See \code{dplyr::\link[dplyr:group_by]{group_by}} for details.
}
\keyword{internal}
