% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{ggp}
\alias{ggp}
\title{ggp}
\usage{
ggp(data = NULL, mapping = aes(), ..., environment = parent.frame())
}
\description{
See \code{ggplot2::\link[ggplot2:ggplot]{ggplot}} for details.
}
\keyword{internal}
