% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{ghl}
\alias{ghl}
\title{ghl}
\usage{
ghl(
  mapping = NULL,
  data = NULL,
  ...,
  yintercept,
  na.rm = FALSE,
  show.legend = NA
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_hline]{geom_hline}} for details.
}
\keyword{internal}
