% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{glb}
\alias{glb}
\title{glb}
\usage{
glb(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  label.padding = unit(0.25, "lines"),
  label.r = unit(0.15, "lines"),
  label.size = 0.25,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_label]{geom_label}} for details.
}
\keyword{internal}
