% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{gln}
\alias{gln}
\title{gln}
\usage{
gln(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_line]{geom_line}} for details.
}
\keyword{internal}
