% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{glu}
\alias{glu}
\title{glu}
\usage{
glu(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .transformer = identity_transformer,
  .trim = TRUE
)
}
\description{
See \code{glue::\link[glue:glue]{glue}} for details.
}
\keyword{internal}
