% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{gpl}
\alias{gpl}
\title{gpl}
\usage{
gpl(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  rule = "evenodd",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_polygon]{geom_polygon}} for details.
}
\keyword{internal}
