% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{gsg}
\alias{gsg}
\title{gsg}
\usage{
gsg(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_segment]{geom_segment}} for details.
}
\keyword{internal}
