% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{gtx}
\alias{gtx}
\title{gtx}
\usage{
gtx(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_text]{geom_text}} for details.
}
\keyword{internal}
