% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{gvl}
\alias{gvl}
\title{gvl}
\usage{
gvl(
  mapping = NULL,
  data = NULL,
  ...,
  xintercept,
  na.rm = FALSE,
  show.legend = NA
)
}
\description{
See \code{ggplot2::\link[ggplot2:geom_vline]{geom_vline}} for details.
}
\keyword{internal}
