% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{hread}
\alias{hread}
\title{hread}
\usage{
hread(x, encoding = "", ..., options = c("RECOVER", "NOERROR",
  "NOBLANKS"))
}
\description{
See \code{xml2::\link[xml2:read_html]{read_html}} for details.
}
\keyword{internal}
