% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{htable}
\alias{htable}
\title{htable}
\usage{
htable(x, header = NA, trim = TRUE, fill = FALSE, dec = ".")
}
\description{
See \code{rvest::\link[rvest:html_table]{html_table}} for details.
}
\keyword{internal}
