% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{labs}
\alias{labs}
\title{labs}
\usage{
labs(
  ...,
  title = waiver(),
  subtitle = waiver(),
  caption = waiver(),
  tag = waiver()
)
}
\description{
See \code{ggplot2::\link[ggplot2:labs]{labs}} for details.
}
\keyword{internal}
