% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.R
\name{pj}
\alias{pj}
\title{Parse json}
\usage{
pj(x)
}
\arguments{
\item{x}{Input object. If a character vector is supplied then JSON content
is parsed from the text. If a JSON file, URL, or xml_document, then the
content is extracted into a character vector and then JSON content parsed.}
}
\value{
A list representing the JSON data
}
\description{
Retrieves and/or parses JSON inputs into R lists and such
}
