% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{sfvc}
\alias{sfvc}
\title{sfvc}
\usage{
sfvc(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill"
)
}
\description{
See \code{ggplot2::\link[ggplot2:scale_fill_viridis_c]{scale_fill_viridis_c}} for details.
}
\keyword{internal}
