% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{sfvd}
\alias{sfvd}
\title{sfvd}
\usage{
sfvd(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  aesthetics = "fill"
)
}
\description{
See \code{ggplot2::\link[ggplot2:scale_fill_viridis_d]{scale_fill_viridis_d}} for details.
}
\keyword{internal}
