% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{spr}
\alias{spr}
\title{spr}
\usage{
spr(data, key, value, fill = NA, convert = FALSE, drop = TRUE, sep = NULL)
}
\description{
See \code{tidyr::\link[tidyr:spread]{spread}} for details.
}
\keyword{internal}
