% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{ssm}
\alias{ssm}
\title{ssm}
\usage{
ssm(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  method = NULL,
  formula = NULL,
  se = TRUE,
  n = 80,
  span = 0.75,
  fullrange = FALSE,
  level = 0.95,
  method.args = list(),
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
See \code{ggplot2::\link[ggplot2:stat_smooth]{stat_smooth}} for details.
}
\keyword{internal}
