% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{sxc}
\alias{sxc}
\title{sxc}
\usage{
sxc(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  n.breaks = NULL,
  labels = waiver(),
  limits = NULL,
  expand = waiver(),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  guide = waiver(),
  position = "left",
  sec.axis = waiver()
)

sxc(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  n.breaks = NULL,
  labels = waiver(),
  limits = NULL,
  expand = waiver(),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  guide = waiver(),
  position = "left",
  sec.axis = waiver()
)
}
\description{
See \code{ggplot2::\link[ggplot2:scale_x_continuous]{scale_x_continuous}} for details.

See \code{ggplot2::\link[ggplot2:scale_y_continuous]{scale_y_continuous}} for details.
}
\keyword{internal}
