% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{sxd}
\alias{sxd}
\title{sxd}
\usage{
sxd(..., expand = waiver(), guide = waiver(), position = "bottom")
}
\description{
See \code{ggplot2::\link[ggplot2:scale_x_discrete]{scale_x_discrete}} for details.
}
\keyword{internal}
