% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ggplot.R
\name{syd}
\alias{syd}
\title{syd}
\usage{
syd(..., expand = waiver(), guide = waiver(), position = "left")
}
\description{
See \code{ggplot2::\link[ggplot2:scale_y_discrete]{scale_y_discrete}} for details.
}
\keyword{internal}
