% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{unn}
\alias{unn}
\title{unn}
\usage{
unn(data, cols, ..., keep_empty = FALSE, ptype = NULL,
  names_sep = NULL, names_repair = "check_unique",
  .drop = deprecated(), .id = deprecated(), .sep = deprecated(),
  .preserve = deprecated())
}
\description{
See \code{tidyr::\link[tidyr:unnest]{unnest}} for details.
}
\keyword{internal}
