% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabb.R
\name{tabb}
\alias{tabb}
\title{tabb}
\usage{
tabb(..., sort = TRUE)
}
\arguments{
\item{...}{One or more vectors, a data frame, or a data frame with tidy-
selected columns}

\item{sort}{Logical indicating whether to sort in order from most to least
frequent, defaults to TRUE}
}
\description{
Generates a frequency table tibble
}
