% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-funs.R
\name{unn}
\alias{unn}
\title{unn}
\usage{
unn(data, ..., .drop = NA, .id = NULL, .sep = NULL,
  .preserve = NULL)
}
\description{
See \code{tidyr::\link[tidyr:unnest]{unnest}} for details.
}
\keyword{internal}
